+++ src/main/java/com/dart/api/domain/chat/repository/ChatRedisRepository.java
@@ -1,6 +1,6 @@
 package com.dart.api.domain.chat.repository;
 
-import static com.dart.api.infrastructure.redis.RedisConstant.*;
+import static com.dart.global.common.util.RedisConstant.*;
 
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
